requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
opencv-python-headless
